# backend/swagger_server/swagger/swagger.yaml
openapi: 3.0.1
info:
  title: Bumpy API
  description: This is the amazing Bumpy API that powers both our web interface and
    the Android app.
  version: 1.0.0
servers:
- url: http://localhost/v1
tags:
- name: trip
  description: Everything about trips
- name: mapData
  description: Operations related to retrieving data to be shown on a map
- name: device
  description: Operations about device
paths:
  /trip/uploadMotionFile:
    post:
      tags:
      - trip
      summary: Uploads a csv motion file, only a few checks are performed.
      operationId: upload_motion_file
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/body'
        required: true
      responses:
        "200":
          description: Upload successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "405":
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "500":
          description: Unexpected exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      x-openapi-router-controller: swagger_server.controllers.trip_controller
  /trip/insertNewTrip:
    post:
      tags:
      - trip
      summary: Insert a new bike trip and start the background processing.
      operationId: insert_new_trip
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Trip'
        required: true
      responses:
        "200":
          description: Validation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "405":
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "500":
          description: Unexpected exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      x-openapi-router-controller: swagger_server.controllers.trip_controller
  /trip/getTripsByDeviceUUID:
    get:
      tags:
      - trip
      summary: Gets all the trips given a deviceUUID
      operationId: get_trips_by_device_uuid
      parameters:
      - name: deviceUUID
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProcessedTrip'
                x-content-type: application/json
        "500":
          description: Unexpected exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      x-openapi-router-controller: swagger_server.controllers.trip_controller
  /trip/getTripByTripUUID:
    get:
      tags:
      - trip
      summary: Gets a trip given a tripUUID
      operationId: get_trip_by_trip_uuid
      parameters:
      - name: tripUUID
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FullProcessedTrip'
        "404":
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "500":
          description: Unexpected exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      x-openapi-router-controller: swagger_server.controllers.trip_controller
  /mapData/getRoadQualitySegments:
    get:
      tags:
      - mapData
      summary: Get the paths composed of segments to display the heatmap
      operationId: get_road_quality_segments
      parameters:
      - name: bottomLeftLat
        in: query
        description: The bottom left coordinate of the corner of the screen
        required: true
        style: form
        explode: true
        schema:
          maximum: 90
          minimum: 0
          type: number
      - name: bottomLeftLon
        in: query
        description: The bottom left coordinate of the corner of the screen
        required: true
        style: form
        explode: true
        schema:
          maximum: 180
          minimum: 0
          type: number
      - name: topRightLat
        in: query
        description: The top right coordinate of the corner of the screen
        required: true
        style: form
        explode: true
        schema:
          maximum: 90
          minimum: 0
          type: number
      - name: topRightLon
        in: query
        description: The top right coordinate of the corner of the screen
        required: true
        style: form
        explode: true
        schema:
          maximum: 180
          minimum: 0
          type: number
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Path'
                x-content-type: application/json
        "500":
          description: Unexpected exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      x-openapi-router-controller: swagger_server.controllers.map_data_controller
  /mapData/getBumpyIssuePoints:
    get:
      tags:
      - mapData
      summary: Get the autocomputed points
      operationId: get_bumpy_issue_points
      parameters:
      - name: bottomLeftLat
        in: query
        description: The bottom left coordinate of the corner of the screen
        required: true
        style: form
        explode: true
        schema:
          maximum: 90
          minimum: 0
          type: number
      - name: bottomLeftLon
        in: query
        description: The bottom left coordinate of the corner of the screen
        required: true
        style: form
        explode: true
        schema:
          maximum: 180
          minimum: 0
          type: number
      - name: topRightLat
        in: query
        description: The top right coordinate of the corner of the screen
        required: true
        style: form
        explode: true
        schema:
          maximum: 90
          minimum: 0
          type: number
      - name: topRightLon
        in: query
        description: The top right coordinate of the corner of the screen
        required: true
        style: form
        explode: true
        schema:
          maximum: 180
          minimum: 0
          type: number
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BumpyPoint'
                x-content-type: application/json
        "500":
          description: Unexpected exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      x-openapi-router-controller: swagger_server.controllers.map_data_controller
  /mapData/getFixMyStreetIssuePoints:
    get:
      tags:
      - mapData
      summary: Get the FixMyStreet points
      operationId: get_fix_my_street_issue_points
      parameters:
      - name: bottomLeftLat
        in: query
        description: The bottom left coordinate of the corner of the screen
        required: true
        style: form
        explode: true
        schema:
          maximum: 90
          minimum: 0
          type: number
      - name: bottomLeftLon
        in: query
        description: The bottom left coordinate of the corner of the screen
        required: true
        style: form
        explode: true
        schema:
          maximum: 180
          minimum: 0
          type: number
      - name: topRightLat
        in: query
        description: The top right coordinate of the corner of the screen
        required: true
        style: form
        explode: true
        schema:
          maximum: 90
          minimum: 0
          type: number
      - name: topRightLon
        in: query
        description: The top right coordinate of the corner of the screen
        required: true
        style: form
        explode: true
        schema:
          maximum: 180
          minimum: 0
          type: number
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FixMyStreetPoint'
                x-content-type: application/json
        "500":
          description: Unexpected exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      x-openapi-router-controller: swagger_server.controllers.map_data_controller
  /device/getShortDeviceUUID:
    get:
      tags:
      - device
      summary: Get the short device UUID from the long device UUID
      operationId: get_short_device_uuid
      parameters:
      - name: deviceUUID
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                maxLength: 6
                minLength: 6
                type: string
                x-content-type: application/json
        "500":
          description: Unexpected exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      x-openapi-router-controller: swagger_server.controllers.device_controller
  /device/getLongDeviceUUID:
    get:
      tags:
      - device
      summary: Get the long device UUID from the short device UUID
      operationId: get_long_device_uuid
      parameters:
      - name: shortDeviceUUID
        in: query
        required: true
        style: form
        explode: true
        schema:
          maxLength: 6
          minLength: 6
          type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: string
                format: uuid
                x-content-type: application/json
        "404":
          description: Device UUID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "500":
          description: Unexpected exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      x-openapi-router-controller: swagger_server.controllers.device_controller
components:
  schemas:
    Trip:
      type: object
      properties:
        deviceUUID:
          type: string
          format: uuid
        tripUUID:
          type: string
          format: uuid
        startTS:
          type: string
          format: date-time
        endTS:
          type: string
          format: date-time
        gnssData:
          type: array
          items:
            $ref: '#/components/schemas/Trip_gnssData'
    ProcessedTrip:
      type: object
      properties:
        deviceUUID:
          type: string
          format: uuid
        tripUUID:
          type: string
          format: uuid
        startTS:
          type: string
          format: date-time
        endTS:
          type: string
          format: date-time
        distance:
          minimum: 0
          type: number
          description: Expressed in meters
        speed:
          $ref: '#/components/schemas/ProcessedTrip_speed'
        elevation:
          $ref: '#/components/schemas/ProcessedTrip_elevation'
      example:
        elevation:
          minElevation: 5.637376656633329
          maxElevation: 5.962133916683182
          avgElevation: 2.3021358869347655
        distance: 0.08008281904610115
        endTS: 2000-01-23T04:56:07.000+00:00
        startTS: 2000-01-23T04:56:07.000+00:00
        tripUUID: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        deviceUUID: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        speed:
          maxSpeed: 0.6027456183070403
          avgSpeed: 0.14658129805029452
    FullProcessedTrip:
      type: object
      properties:
        deviceUUID:
          type: string
          format: uuid
        tripUUID:
          type: string
          format: uuid
        startTS:
          type: string
          format: date-time
        endTS:
          type: string
          format: date-time
        distance:
          minimum: 0
          type: number
          description: Expressed in meters
        speed:
          $ref: '#/components/schemas/ProcessedTrip_speed'
        elevation:
          $ref: '#/components/schemas/ProcessedTrip_elevation'
        gnssData:
          type: array
          items:
            $ref: '#/components/schemas/Trip_gnssData'
      example:
        elevation:
          minElevation: 5.637376656633329
          maxElevation: 5.962133916683182
          avgElevation: 2.3021358869347655
        distance: 0.08008281904610115
        endTS: 2000-01-23T04:56:07.000+00:00
        gnssData:
        - timeTS: 2000-01-23T04:56:07.000+00:00
          precision: 5.962133916683182
          lon: 26.384633649053015
          lat: 54.24710564763363
          speed: 2.3021358869347655
          ele: 5.637376656633329
        - timeTS: 2000-01-23T04:56:07.000+00:00
          precision: 5.962133916683182
          lon: 26.384633649053015
          lat: 54.24710564763363
          speed: 2.3021358869347655
          ele: 5.637376656633329
        startTS: 2000-01-23T04:56:07.000+00:00
        tripUUID: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        deviceUUID: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        speed:
          maxSpeed: 0.6027456183070403
          avgSpeed: 0.14658129805029452
    Path:
      type: object
      properties:
        pathUUID:
          type: string
          format: uuid
        segments:
          type: array
          items:
            $ref: '#/components/schemas/Path_segments'
      example:
        pathUUID: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        segments:
        - startLat: 7.2074537141491035
          endLat: 13.192316824526507
          startLon: 108.49421129526726
          endLon: 107.31841050029728
          qualityScore: 5.637376656633329
          segmentUUID: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        - startLat: 7.2074537141491035
          endLat: 13.192316824526507
          startLon: 108.49421129526726
          endLon: 107.31841050029728
          qualityScore: 5.637376656633329
          segmentUUID: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
    BumpyPoint:
      type: object
      properties:
        pointUUID:
          type: string
          format: uuid
        lat:
          maximum: 90
          minimum: 0
          type: number
        lon:
          maximum: 180
          minimum: 0
          type: number
        bumpScore:
          type: number
      example:
        lon: 108.49421129526726
        pointUUID: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        lat: 7.2074537141491035
        bumpScore: 1.4658129805029452
    FixMyStreetPoint:
      type: object
      properties:
        pointUUID:
          type: string
          format: uuid
        lat:
          maximum: 90
          minimum: 0
          type: number
        lon:
          maximum: 180
          minimum: 0
          type: number
        fixMyStreetUrl:
          type: string
          format: url
      example:
        fixMyStreetUrl: http://example.com/aeiou
        lon: 108.49421129526726
        pointUUID: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        lat: 7.2074537141491035
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
        type:
          type: string
        message:
          type: string
      example:
        code: 0
        type: type
        message: message
    body:
      type: object
      properties:
        file:
          type: string
          format: binary
        tripUUID:
          type: string
          format: uuid
    Trip_gnssData:
      type: object
      properties:
        timeTS:
          type: string
          format: date-time
        lat:
          maximum: 90
          minimum: 0
          type: number
        lon:
          maximum: 180
          minimum: 0
          type: number
        precision:
          type: number
          description: Expressed in meters
        ele:
          type: number
          description: Expressed in meters
        speed:
          type: number
          description: Expressed in meters per second
      example:
        timeTS: 2000-01-23T04:56:07.000+00:00
        precision: 5.962133916683182
        lon: 26.384633649053015
        lat: 54.24710564763363
        speed: 2.3021358869347655
        ele: 5.637376656633329
    Path_segments:
      type: object
      properties:
        segmentUUID:
          type: string
          format: uuid
        startLat:
          maximum: 90
          minimum: 0
          type: number
        startLon:
          maximum: 180
          minimum: 0
          type: number
        endLat:
          maximum: 90
          minimum: 0
          type: number
        endLon:
          maximum: 180
          minimum: 0
          type: number
        qualityScore:
          type: number
      example:
        startLat: 7.2074537141491035
        endLat: 13.192316824526507
        startLon: 108.49421129526726
        endLon: 107.31841050029728
        qualityScore: 5.637376656633329
        segmentUUID: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
    ProcessedTrip_speed:
      type: object
      properties:
        maxSpeed:
          minimum: 0
          type: number
          description: Expressed in meters per second
        avgSpeed:
          minimum: 0
          type: number
          description: Expressed in meters per second
      example:
        maxSpeed: 0.6027456183070403
        avgSpeed: 0.14658129805029452
    ProcessedTrip_elevation:
      type: object
      properties:
        maxElevation:
          type: number
          description: Expressed in meters
        minElevation:
          type: number
          description: Expressed in meters
        avgElevation:
          type: number
          description: Expressed in meters
      example:
        minElevation: 5.637376656633329
        maxElevation: 5.962133916683182
        avgElevation: 2.3021358869347655